Fix errors when QCOW2 image is spawned on VMWare
When QCOW2 image is spawned on VMWare, we first create port-group and
allocate VLAN first and then call spawn. If the spawn API fails, we need
to,
- Delete port-group
- Free VLAN
- Call DeletePort() to agent

When VM spawn fails, nova invokes destroy() method in contrail.py. But,
the network-info on destroy does not contrail all necessary information.
So, we are storing all instance related parameters in vm_info object and
using vm_info for all cleanup.

Change-Id: Icb2367405a6b062b16a2c6f090b8c58166616ab7